The 6:50 AM casting combat patch

+++ src/boot/MiscUtils.h
@@ -148,11 +148,12 @@ EFI_STATUS BlDecompressLZVN(VOID CONST* compressedBuffer, UINTN compressedSize, 
 #define CASE_TABLE	127
 
 #define OSSwapInt64(x) \
-((UINT64)((((UINT64)(x) & 0xff00000000000000ULL) >> 56) | \
-(((UINT64)(x) & 0x00ff000000000000ULL) >> 40) | \
-(((UINT64)(x) & 0x0000ff0000000000ULL) >> 24) | \
-(((UINT64)(x) & 0x000000ff00000000ULL) >>  8) | \
-(((UINT64)(x) & 0x00000000ff000000ULL) <<  8) | \
-(((UINT64)(x) & 0x0000000000ff0000ULL) << 24) | \
-(((UINT64)(x) & 0x000000000000ff00ULL) << 40) | \
-(((UINT64)(x) & 0x00000000000000ffULL) << 56)))
+	((((x) & 0xff00000000000000ULL) >> 56) | \
+	(((x) & 0x00ff000000000000ULL) >> 40) | \
+	(((x) & 0x0000ff0000000000ULL) >> 24) | \
+	(((x) & 0x000000ff00000000ULL) >>  8) | \
+	(((x) & 0x00000000ff000000ULL) <<  8) | \
+	(((x) & 0x0000000000ff0000ULL) << 24) | \
+	(((x) & 0x000000000000ff00ULL) << 40) | \
+	(((x) & 0x00000000000000ffULL) << 56))
+
